+++ src/components/app.test.js
@@ -6,11 +6,11 @@ import App from './app';
 import Introduction from './introduction';
 import BasicTutorial from './tutorials/basic-tutorial';
 import AdvancedTutorial from './tutorials/advanced-tutorial';
-import Submissions from './api-reference/schemas/submissions';
-import MeasurementSets from './api-reference/schemas/measurement-sets';
-import Measurements from './api-reference/schemas/measurements';
-import Benchmarks from './api-reference/schemas/benchmarks';
-import Scoring from './api-reference/scoring';
+import Submissions from './reference/schemas/submissions';
+import MeasurementSets from './reference/schemas/measurement-sets';
+import Measurements from './reference/schemas/measurements';
+import Benchmarks from './reference/schemas/benchmarks';
+import Scoring from './reference/scoring';
 
 it('renders without crashing', () => {
   shallow(<App />);
