+++ src/components/references/common/data-scoring-navigation-table.js
