+++ src/components/references/common/scoring-example-output-aci.json
