+++ src/components/references/common/scoring-example-output-ia.json
