+++ src/components/references/common/scoring-example-submission-input.json
